requirements.txt
tensorflow
werkzeug
pillow
valohai-utils